requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
python-dotenv==1.0.1
websockets==12.0
google-genai>=0.1.0